Fix issues

+++ app/overlay/motion/frame_diff_detector.py
@@ -62,7 +62,7 @@ class FrameDiffDetector(BaseMotionDetector):
         height, width = gray.shape
 
         # Проверка brightness (edge case)
-        avg_brightness = gray.mean()
+        avg_brightness = float(gray.mean())
         if avg_brightness < self.min_brightness:
             return np.zeros((self.grid_height, self.grid_width), dtype=np.uint8)
